add macro for easy forward declaration
Summary:
## Why would we want to forward declare my_enum::enums?

Because my_enum's enum are not only enum but comes with a lot of helpers functions,
it turns out including the my_enums real definition is surprisingly expensive at compile time.

This can be seen for long enum like the one in:

`arvr/libraries/perception/calib_structs/Common.h`

which takes 1 second to parse (roughly equivalent to parsing `Eigen/Dense`!)

Sometime we just need to forward declare the enum and we can avoid those slow header creeping the entire include hiearchy, with the potential of saving a 1-2 seconds of turnaround time when iterating on code locally. It won't save much in general.

Reviewed By: h-friederich

Differential Revision: D68600259

fbshipit-source-id: 18b58f583cf2105610b49f2ac539fa338991454d