add `tryFromString` sugar
Summary:
The usage pattern with the pre-existing `trySetFromString` is a bit annoying.
You have to first declare the Enum variable, then try to set it in a different line.
See e.g.
https://www.internalfb.com/code/fbsource/[77bf31292bb1]/arvr/projects/xrtech/engines/InsideOutTracking/InsideOutTrackingEngine.cpp?lines=2173-2174

Returning Optional has nicer semantics and allows for a one-liner.
Since this is a 3p library we cannot depend on perception::Optional though.

Here I create the `tryFromString` version for both case-sensitive and case-insensitive `trySetFromString` methods.

Reviewed By: h-friederich

Differential Revision: D72504230

fbshipit-source-id: 02716b436669a3a8898d934e7fad9ec6b4d03ae8